+++ CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## 2.0.2 - 2018-02-20
+## 2.1.0 - 2019-03-18
+### Added
+- [Blitz ⚡️](https://plugins.craftcms.com/blitz) support! Updating your navigations will now also clear & warm Blitz's cache if it's installed (requires Blitz 2.0.1)
+- Navigation caches can now be cleared from the CP, under utilities/clear-caches.
+
+## 2.0.2 - 2019-02-20
 ### Fixed
 - Fixed an issue where a disabled element would still be included in the frontend query ([#21](https://github.com/studioespresso/craft3-navigate/issues/21))
 - Fixed an issue where editing a navigaation would always link to the default site navigation ([#23](https://github.com/studioespresso/craft3-navigate/issues/23))
@@ -14,11 +19,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - The site selection dropdown is no longer shown in the entries/assets/categories modal ([#23](https://github.com/studioespresso/craft3-navigate/issues/23))
 - URL nodes are now also parsed for variables from your .env file
 
-## 2.0.1 - 2018-01-30
+## 2.0.1 - 2019-01-30
 ### Fixed
 - Fixed an issue where disabled child nodes would be included in site querries
 
-## 2.0.0 - 2018-01-16
+## 2.0.0 - 2019-01-16
 ### Added
 - Craft 3.1 is here! Navigate is fully compatible and now support Craft's "project config", allowing you to include navigations in your project.yaml